Catcher/CFG_Project/requirements.txt
mysql-connector-python==8.0.27
pytest==6.2.5
pygame~=2.0.0.dev6